move mcts options building into separate function

+++ cpp/common/inl/Mcts.inl
@@ -16,17 +16,55 @@
 namespace common {
 
 template<GameStateConcept GameState, TensorizorConcept<GameState> Tensorizor>
-boost::filesystem::path Mcts_<GameState, Tensorizor>::kProfilingDir;
+typename Mcts_<GameState, Tensorizor>::Params Mcts_<GameState, Tensorizor>::global_params_;
+
+template<GameStateConcept GameState, TensorizorConcept<GameState> Tensorizor>
+void Mcts_<GameState, Tensorizor>::add_options(boost::program_options::options_description& desc) {
+  namespace po = boost::program_options;
+
+  boost::filesystem::path default_nnet_filename_path = util::Repo::root() / "c4_model.pt";
+  std::string default_nnet_filename = util::Config::instance()->get(
+      "nnet_filename", default_nnet_filename_path.string());
+
+  boost::filesystem::path default_profiling_dir_path = util::Repo::root() / "output" / "mcts_profiling";
+  std::string default_profiling_dir = util::Config::instance()->get(
+      "mcts_profiling_dir", default_profiling_dir_path.string());
+
+  Params& params = global_params_;
+  desc.add_options()
+      ("mcts-nnet-filename", po::value<std::string>(&params.nnet_filename)->default_value(default_nnet_filename),
+          "nnet filename")
+      ("mcts-num-search-threads", po::value<int>(&params.num_search_threads)->default_value(params.num_search_threads),
+          "num search threads")
+      ("mcts-batch-size-limit", po::value<int>(&params.batch_size_limit)->default_value(params.batch_size_limit),
+          "batch size limit")
+      ("mcts-run-offline", po::bool_switch(&params.run_offline)->default_value(false),
+          "run search while opponent is thinking")
+      ("mcts-offline-tree-size-limit", po::value<int>(&params.offline_tree_size_limit)->default_value(
+          params.offline_tree_size_limit), "max tree size to grow to offline (only respected in --run-offline mode)")
+      ("mcts-nn-eval-timeout-ns", po::value<int64_t>(&params.nn_eval_timeout_ns)->default_value(
+          params.nn_eval_timeout_ns), "nn eval thread timeout in ns")
+      ("mcts-cache-size", po::value<size_t>(&params.cache_size)->default_value(params.cache_size),
+          "nn eval thread cache size")
+      ("mcts-root-softmax-temp", po::value<float>(&params.root_softmax_temperature)->default_value(
+          params.root_softmax_temperature), "root softmax temperature")
+      ("mcts-cpuct", po::value<float>(&params.cPUCT)->default_value(params.cPUCT), "cPUCT value")
+      ("mcts-dirichlet-mult", po::value<float>(&params.dirichlet_mult)->default_value(params.dirichlet_mult),
+          "dirichlet mult")
+      ("mcts-dirichlet-alpha", po::value<float>(&params.dirichlet_alpha)->default_value(params.dirichlet_alpha),
+          "dirichlet alpha")
+      ("mcts-allow-eliminations", po::bool_switch(&params.allow_eliminations)->default_value(
+          params.allow_eliminations), "allow eliminations")
+#ifdef PROFILE_MCTS
+      ("mcts-profiling-dir", po::value<std::string>(&params.profiling_dir)->default_value(default_profiling_dir),
+          "directory in which to dump mcts profiling stats")
+#endif  // PROFILE_MCTS
+      ;
+}
 
 template<GameStateConcept GameState, TensorizorConcept<GameState> Tensorizor>
 typename Mcts_<GameState, Tensorizor>::NNEvaluationService::instance_map_t
 Mcts_<GameState, Tensorizor>::NNEvaluationService::instance_map_;
-
-template<GameStateConcept GameState, TensorizorConcept<GameState> Tensorizor>
-inline boost::filesystem::path Mcts_<GameState, Tensorizor>::default_profiling_dir() {
-  boost::filesystem::path default_dir = util::Repo::root() / "output" / "mcts_profiling";
-  return util::Config::instance()->get("mcts_profiling_dir", default_dir.string());
-}
 
 template<GameStateConcept GameState, TensorizorConcept<GameState> Tensorizor>
 inline Mcts_<GameState, Tensorizor>::NNEvaluation::NNEvaluation(
@@ -279,7 +317,7 @@ template<GameStateConcept GameState, TensorizorConcept<GameState> Tensorizor>
 inline Mcts_<GameState, Tensorizor>::SearchThread::SearchThread(Mcts_* mcts, int thread_id)
 : mcts_(mcts)
 , thread_id_(thread_id) {
-  auto profiling_filename = kProfilingDir / util::create_string("search%d.txt", thread_id);
+  auto profiling_filename = mcts->profiling_dir() / util::create_string("search%d.txt", thread_id);
   set_profiling_file(profiling_filename.c_str());
 }
 
@@ -358,15 +396,17 @@ inline void Mcts_<GameState, Tensorizor>::SearchThread::profiling_stats_t::clear
 
 template<GameStateConcept GameState, TensorizorConcept<GameState> Tensorizor>
 typename Mcts_<GameState, Tensorizor>::NNEvaluationService*
-Mcts_<GameState, Tensorizor>::NNEvaluationService::create(
-    const boost::filesystem::path& net_filename, int batch_size_limit, int64_t timeout_ns,
-    size_t cache_size)
-{
+Mcts_<GameState, Tensorizor>::NNEvaluationService::create(const Mcts_* mcts) {
+  int64_t timeout_ns = mcts->params().nn_eval_timeout_ns;
+  boost::filesystem::path net_filename(mcts->params().nnet_filename);
+  size_t cache_size = mcts->params().cache_size;
+  int batch_size_limit = mcts->params().batch_size_limit;
+
   std::chrono::nanoseconds timeout_duration(timeout_ns);
   auto it = instance_map_.find(net_filename);
   if (it == instance_map_.end()) {
     NNEvaluationService* instance = new NNEvaluationService(
-        net_filename, batch_size_limit, timeout_duration, cache_size);
+        net_filename, batch_size_limit, timeout_duration, cache_size, mcts->profiling_dir());
     instance_map_[net_filename] = instance;
     return instance;
   }
@@ -406,7 +446,7 @@ void Mcts_<GameState, Tensorizor>::NNEvaluationService::disconnect() {
 template<GameStateConcept GameState, TensorizorConcept<GameState> Tensorizor>
 inline Mcts_<GameState, Tensorizor>::NNEvaluationService::NNEvaluationService(
     const boost::filesystem::path& net_filename, int batch_size, std::chrono::nanoseconds timeout_duration,
-    size_t cache_size)
+    size_t cache_size, const boost::filesystem::path& profiling_dir)
 : net_(net_filename)
 , policy_batch_(batch_size, kNumGlobalActions, util::to_std_array<int>(batch_size, kNumGlobalActions))
 , value_batch_(batch_size, kNumPlayers, util::to_std_array<int>(batch_size, kNumPlayers))
@@ -421,8 +461,10 @@ inline Mcts_<GameState, Tensorizor>::NNEvaluationService::NNEvaluationService(
   input_vec_.push_back(torch_input_gpu_);
   deadline_ = std::chrono::steady_clock::now();
 
-  auto profiling_filename = kProfilingDir / "eval.txt";
-  set_profiling_file(profiling_filename.c_str());
+  if (kEnableProfiling) {
+    auto profiling_filename = profiling_dir / "eval.txt";
+    set_profiling_file(profiling_filename.c_str());
+  }
 }
 
 template<GameStateConcept GameState, TensorizorConcept<GameState> Tensorizor>
@@ -601,8 +643,9 @@ void Mcts_<GameState, Tensorizor>::NNEvaluationService::record_for_profiling(reg
 template<GameStateConcept GameState, TensorizorConcept<GameState> Tensorizor>
 inline Mcts_<GameState, Tensorizor>::Mcts_(const Params& params)
 : params_(params) {
-  nn_eval_service_ = NNEvaluationService::create(
-      params.nnet_filename, params.batch_size_limit, params.nn_eval_timeout_ns, params.cache_size);
+  namespace bf = boost::filesystem;
+
+  nn_eval_service_ = NNEvaluationService::create(this);
   if (num_search_threads() < 1) {
     throw util::Exception("num_search_threads must be positive (%d)", num_search_threads());
   }
@@ -614,6 +657,13 @@ inline Mcts_<GameState, Tensorizor>::Mcts_(const Params& params)
   }
   for (int i = 0; i < num_search_threads(); ++i) {
     search_threads_.push_back(new SearchThread(this, i));
+  }
+
+  if (kEnableProfiling) {
+    if (profiling_dir().empty()) {
+      throw util::Exception("Required: --mcts-profiling-dir. Alternatively, add entry for 'mcts_profiling_dir' in config.txt");
+    }
+    init_profiling_dir(profiling_dir().string());
   }
 }
 
@@ -669,7 +719,7 @@ inline void Mcts_<GameState, Tensorizor>::receive_state_change(
   root_->_adopt_children();
 
   if (params_.run_offline) {
-    start_search_threads(params_.max_tree_size_limit);
+    start_search_threads(params_.offline_tree_size_limit);
   }
 }
 
@@ -842,13 +892,20 @@ inline void Mcts_<GameState, Tensorizor>::run_search(SearchThread* thread, int t
 }
 
 template<GameStateConcept GameState, TensorizorConcept<GameState> Tensorizor>
-void Mcts_<GameState, Tensorizor>::set_profiling_dir(const boost::filesystem::path& path) {
+void Mcts_<GameState, Tensorizor>::init_profiling_dir(const std::string& profiling_dir) {
+  static std::string pdir;
+  if (!pdir.empty()) {
+    if (pdir == profiling_dir) return;
+    throw util::Exception("Two different mcts profiling dirs used: %s and %s", pdir.c_str(), profiling_dir.c_str());
+  }
+  pdir = profiling_dir;
+
   namespace bf = boost::filesystem;
+  bf::path path(profiling_dir);
   if (bf::is_directory(path)) {
     bf::remove_all(path);
   }
   bf::create_directories(path);
-  kProfilingDir = path;
 }
 
 template<GameStateConcept GameState, TensorizorConcept<GameState> Tensorizor>
